refactor: перенести KogdaIgraGameSelector в JoinRpg.Web.ProjectCommon и добавить список будущих игр
Клиентские интерфейсы/DTO КогдаИгры (IKogdaIgraSyncClient, IKogdaIgraBindClient,
KogdaIgraShortViewModel, KogdaIgraCardViewModel и т.д.) переехали из
JoinRpg.Web.AdminTools в JoinRpg.Web.ProjectCommon, чтобы KogdaIgraGameSelector
стал доступен не только админке. Компонент получил параметры Multiple
(одиночный/множественный выбор) и FutureOnly (загружать только будущие игры) —
дефолты сохраняют текущее поведение админских экранов без изменений.

Добавлены IKogdaIgraRepository.GetActiveFuture() и публичный (не только для
админов) эндпоинт GetFutureKogdaIgraCandidates — понадобятся форме создания
проекта.
