+++ src/JoinRpg.Data.Interfaces/AdminTools/IKogdaIgraRepository.cs
@@ -8,6 +8,7 @@ public record KogdaIgraListItem(KogdaIgraIdentification KogdaIgraId, string Name
 public interface IKogdaIgraRepository
 {
     Task<KogdaIgraListItem[]> GetActive();
+    Task<KogdaIgraListItem[]> GetActiveFuture();
     Task<ICollection<KogdaIgraGame>> GetByIds(IReadOnlyCollection<KogdaIgraIdentification> kogdaIgraIdentifications);
 
     Task<IReadOnlyCollection<KogdaIgraGameData>> GetDataByIds(IReadOnlyCollection<KogdaIgraIdentification> kogdaIgraIdentifications);
