+++ src/JoinRpg.Portal/Pages/Admin/SyncWithKogdaIgra.cshtml
@@ -1,5 +1,6 @@
 ﻿@page "/admin/kogda-igra-sync"
 @using JoinRpg.Web.AdminTools.KogdaIgra
+@using JoinRpg.Web.ProjectCommon.KogdaIgra
 @model JoinRpg.Portal.Pages.Admin.SyncWithKogdaIgraModel
 
 @{
