+++ src/JoinRpg.Web.AdminTools/AdminHotRoleViewModel.cs
@@ -1,5 +1,5 @@
-using JoinRpg.Web.Games.Projects;
 using JoinRpg.Web.ProjectCommon;
+using JoinRpg.Web.ProjectCommon.KogdaIgra;
 using JoinRpg.Web.ProjectCommon.Projects;
 using Microsoft.AspNetCore.Components;
