+++ src/JoinRpg.Web.AdminTools/AdminProjectGrid.razor
@@ -1,4 +1,4 @@
-@using JoinRpg.Web.AdminTools.KogdaIgra
+@using JoinRpg.Web.ProjectCommon.KogdaIgra
 @using JoinRpg.Web.ProjectCommon.Projects
 @using System.ComponentModel
 @using System.ComponentModel.DataAnnotations
