+++ src/JoinRpg.Web.AdminTools/KogdaIgra/KogdaIgraForceGameSync.razor
@@ -1,3 +1,4 @@
+@using JoinRpg.Web.ProjectCommon.KogdaIgra
 @inject IKogdaIgraSyncClient Client
 
 <JoinPanel>
