+++ src/JoinRpg.Web.AdminTools/KogdaIgra/KogdaIgraSyncList.razor
@@ -1,3 +1,4 @@
+@using JoinRpg.Web.ProjectCommon.KogdaIgra
 @using JoinRpg.Web.ProjectCommon.Projects
 @using Microsoft.JSInterop
 @using System.ComponentModel.DataAnnotations
