+++ src/JoinRpg.Web.AdminTools/KogdaIgra/KogdaIgraUpdater.razor
@@ -1,6 +1,7 @@
 @inject IKogdaIgraSyncClient Client
 @implements IDisposable
 @using JoinRpg.Web.AdminTools.Jobs
+@using JoinRpg.Web.ProjectCommon.KogdaIgra
 @using Timer = System.Timers.Timer
 
 @if (model is null)
