+++ src/JoinRpg.Web.AdminTools/ProjectAdminControlPanel.razor
@@ -1,4 +1,4 @@
-@using JoinRpg.Web.AdminTools.KogdaIgra
+@using JoinRpg.Web.ProjectCommon.KogdaIgra
 @using JoinRpg.Web.ProjectCommon.Projects
 @using Microsoft.JSInterop
 @using System.ComponentModel.DataAnnotations
