+++ src/JoinRpg.Web.Games/Projects/KogdaIgraCard.razor
@@ -1,3 +1,4 @@
+@using JoinRpg.Web.ProjectCommon.KogdaIgra
 @using JoinRpg.Web.ProjectCommon.Projects
 
 <JoinPanel>
