+++ src/JoinRpg.Web.Games/Projects/KogdaIgraDetails.razor
@@ -1,3 +1,4 @@
+@using JoinRpg.Web.ProjectCommon.KogdaIgra
 @using JoinRpg.Web.ProjectCommon.Projects
 @using JoinRpg.Common.WebComponents
