+++ src/JoinRpg.WebPortal.Managers/AdminTools/ViewModelBuilders.cs
@@ -1,7 +1,7 @@
 using JoinRpg.Common.KogdaIgraClient;
 using JoinRpg.Services.Interfaces.Integrations.KogdaIgra;
-using JoinRpg.Web.AdminTools.KogdaIgra;
 using JoinRpg.Web.Games.Projects;
+using JoinRpg.Web.ProjectCommon.KogdaIgra;
 
 namespace JoinRpg.WebPortal.Managers.AdminTools;
